* Fix issue 430
